antiFraude/requirements.txt
azure.core
azure-ai-documentintelligence
azure-storage-blob
python-dotenv
streamlit 